- flatten abi - move handlers to handler.go
